third_party/blink/web_tests/external/wpt/css/css-transforms/parsing/rotate-parsing-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['rotate'] = "-0.5 0 0 400grad" should set the property value
  assert_equals: serialization should be canonical expected "x -400grad" but got "-0.5 0 0 400grad"
[FAIL] e.style['rotate'] = "-1 0 0 400grad" should set the property value
  assert_equals: serialization should be canonical expected "x -400grad" but got "-1 0 0 400grad"
[FAIL] e.style['rotate'] = "0 -0.5 0 400grad" should set the property value
  assert_equals: serialization should be canonical expected "y -400grad" but got "0 -0.5 0 400grad"
[FAIL] e.style['rotate'] = "0 -1 0 400grad" should set the property value
  assert_equals: serialization should be canonical expected "y -400grad" but got "0 -1 0 400grad"
[FAIL] e.style['rotate'] = "0 0 -0.5 400grad" should set the property value
  assert_equals: serialization should be canonical expected "-400grad" but got "0 0 -0.5 400grad"
[FAIL] e.style['rotate'] = "0 0 -1 400grad" should set the property value
  assert_equals: serialization should be canonical expected "-400grad" but got "0 0 -1 400grad"
Harness: the test ran to completion.
